// services/cogitor-api/src/app/auth/dtos/confirm.dto.ts
export interface ConfirmDto {
  Description?: string;
  RepeatDescription?: string;
  ActivationRequestId?: string;
  ActivationCode?: string;
  SpamWarning?: string;
  Email?: string;
  Login?: string;
  Phone?: string;
  Password?: string;
  RestorePassword?: string;
  FirstName?: string;
  SecondName?: string;
  Surname?: string;
  address?: string;
}